+++ srcs/canvas.c
@@ -6,7 +6,7 @@
 /*   By: chamada <chamada@student.le-101.fr>        +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/12/05 15:04:23 by chamada      #+#   ##    ##    #+#       */
-/*   Updated: 2020/01/12 02:28:52 by chamada     ###    #+. /#+    ###.fr     */
+/*   Updated: 2020/01/19 01:42:30 by chamada     ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -34,7 +34,7 @@ int						init_canvas(t_env *env)
 	return (1);
 }
 
-void					draw_tex(t_env *env, int pos_x, int z, int size)
+void					draw_tex(t_env *env, int pos_x, float z, int size)
 {
 	t_stripe	stripe;
 	t_ivector	start;
